Mitglieder components now take wahlJahr prop (#28)
