+++ frontend/src/components/mitglieder/MitgliederPage.tsx
@@ -2,12 +2,16 @@ import React from "react";
 import {Container} from "react-bootstrap";
 import {MitgliederTable} from "./MitgliederTable";
 
+interface Props {
+    selectedYear: number,
+}
+
 // The "Mitglieder" page
-export class MitgliederPage extends React.Component {
+export class MitgliederPage extends React.Component<Props> {
     render() {
         return (
             <Container>
-                <MitgliederTable/>
+                <MitgliederTable selectedYear={this.props.selectedYear}/>
             </Container>
         )
     }
