+++ frontend/src/components/mitglieder/MitgliederTable.tsx
@@ -4,7 +4,7 @@ import 'react-bootstrap-table-next/dist/react-bootstrap-table2.min.css';
 import MitgliederEndpoints, {Mitglied} from "../../rest_client/MitgliederEndpoints";
 
 interface Props {
-
+    selectedYear: number,
 }
 
 interface State {
@@ -22,7 +22,18 @@ export class MitgliederTable extends React.Component<Props> {
     }
 
     componentDidMount() {
-        MitgliederEndpoints.getAll().then(data => {
+        this.fetchDataAndSetState(this.props.selectedYear);
+    }
+
+    componentDidUpdate(prevProps: Readonly<Props>, prevState: Readonly<{}>, snapshot?: any) {
+        // Re-fetch data if selected year has changed
+        if (prevProps.selectedYear !== this.props.selectedYear) {
+            this.fetchDataAndSetState(this.props.selectedYear);
+        }
+    }
+
+    fetchDataAndSetState(year: number) {
+        MitgliederEndpoints.getAll(year).then(data => {
             // Sort the data alphabetically by first name
             data.sort((a, b) => (a.first_name > b.first_name) ? 1 : -1);
             this.setState({
